+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Bug fixes:
 * [OLMIS-7931](https://openlmis.atlassian.net/browse/OLMIS-7931): Fix imports for layout-ui jenkins job
 
+Improvements:
+* [OLMIS-7954](https://openlmis.atlassian.net/browse/OLMIS-7954): Improved the scalability of local storage
+
 7.2.11 / 2024-04-19
 ==================
 Bug fixes: